Re-structure rten-simd to put ops traits in a submodule
 - Extract the SIMD vector, mask and element type traits out of `vec.rs` and
   into separate modules.
 - Rename `rten_simd/src/vec.rs` -> `ops.rs` as a more appropriate name for the
   functionality in the module.
 - Un-export operations traits from the root of the `rten_simd` crate and export
   the whole `ops` module. This reduces the clutter of the crate root and gives
   imports of ops traits a more meaningful path like `use
   rten_simd::ops::Extend`.
 - Add some module documentation to ops.rs
