+++ rten-vecmath/src/min_max.rs
@@ -1,4 +1,5 @@
-use rten_simd::{Isa, NumOps, Simd, SimdIterable, SimdOp};
+use rten_simd::ops::NumOps;
+use rten_simd::{Isa, Simd, SimdIterable, SimdOp};
 
 /// Compute the minimum and maximum values in a slice of floats.
 pub struct MinMax<'a> {
