+++ rten-vecmath/src/quantize.rs
@@ -1,6 +1,7 @@
 use std::mem::MaybeUninit;
 
-use rten_simd::{FloatOps, Isa, NarrowSaturate, NumOps, SimdOp, SliceWriter};
+use rten_simd::ops::{FloatOps, NarrowSaturate, NumOps};
+use rten_simd::{Isa, SimdOp, SliceWriter};
 
 /// Quantize a slice of `f32` elements to 8-bit integers using the formula:
 ///
@@ -77,7 +78,8 @@ impl<'d> SimdOp for Quantize<'_, 'd, u8> {
 
 #[cfg(test)]
 mod tests {
-    use rten_simd::{Isa, NumOps, SimdOp};
+    use rten_simd::ops::NumOps;
+    use rten_simd::{Isa, SimdOp};
 
     use super::Quantize;
