+++ rten-vecmath/src/softmax.rs
@@ -1,8 +1,9 @@
 use std::mem::MaybeUninit;
 
 use rten_simd::functional::simd_map;
+use rten_simd::ops::{FloatOps, NumOps};
 use rten_simd::span::SrcDest;
-use rten_simd::{FloatOps, Isa, NumOps, SimdIterable, SimdOp, SimdUnaryOp};
+use rten_simd::{Isa, SimdIterable, SimdOp, SimdUnaryOp};
 
 use crate::Exp;
