+++ rten-vecmath/src/tanh.rs
@@ -1,6 +1,7 @@
 #![allow(clippy::excessive_precision)]
 
-use rten_simd::{FloatOps, Isa, NumOps, Simd, SimdUnaryOp};
+use rten_simd::ops::{FloatOps, NumOps};
+use rten_simd::{Isa, Simd, SimdUnaryOp};
 
 use crate::Exp;
